Initial game file implementation.
This implements the basic I/O operations for game data files. I'm not
necessarily happy with all of this yet but it's just a first draft.
Right now game files are always created during construction of a
game_file object. In the future I might remove this because there are
circumstances where an application might want to only open an existing
file.