Add game object.
This adds the core game object to the library. Currently, this includes
a constructor and a destructor that handle file I/O. To support file I/O
a data_file.hpp header is created containing some utility types and
constants. Finally, a test is added that ensures I/O correctness.