Update test build file.
This sets up home directory search tests in the build file. The home
directory test requires a bit more preparation than previous tests
because it needs to access the system environment. As a result, the test
is executed multiple times (the exact number varies depending on the
system) with slightly varied configurations.