Expand the game object.
This adds a variety of features to megatech::ttt::game. Most
importantly, this adds new constructors (one for a new game and one for
loading), a stream insertion operator (for printing to std::ostream
objects), logic for turn taking, and logic to ensure that lock files are
unlocked when a game object throws.

Turn taking is probably the most significant addition. Basically, a game
will proceed by first giving the X (first) player a turn. After X takes
a turn, the game will proceed to update its state. This proceeds by
first checking for a win. Next, the game checks to see if the board is
entirely full. If it is then the game is a draw. Otherwise, the game
allows the O (second) player to take a turn. After O's turn the state is
updated the same way.

Single player games are not yet implemented.