Add strategy objects.
This adds a strategy object that defines the behavior of a computer
player for single player tic-tac-toe. The computer player should play a
perfect game but, because it's difficult to develop a good test for
this, I'm not 100% certain that it does. It plays well enough that I
can't fool it and well enough that it avoids the mistakes in test cases
that I've created.

To support the strategy object, changes are made to state objects where
appropriate (i.e., where they don't involve applying game logic to state
inappropriately).

Test cases for strategies are added and build files are updated
accordingly.

Finally, the take-turn application is changed to use the strategy object
when in single player mode. Single player games function as expected.