Documentation and cleanup.
This adds inline Doxygen documentation for the public interface of
libttt. Additionally, I cleaned up some of the various files while
working on it.

A new target is available through meson that will build HTML
documentation with Doxygen. A configuration file for Doxygen is now
provided in the root of the repository.

The README is updated and most files have been tagged with copyright
information.

Finally, the .gitignore file is updated to ignore documentation
directories and some Valgrind output files.